#32 Moved around some of the files and cleaned old and deprecated files from first draft
