+++ src/javascript/browserApi/storageProvider.js
